application/requirements.txt
streamlit~=1.33.0
streamlit-authenticator~=0.3.2
boto3~=1.33.4
psycopg2-binary==2.9.9
SQLAlchemy==1.4.52
opensearch-py==2.4.2
PyMySQL==1.1.1
python-dotenv~=1.0.0
plotly~=5.18.0
cryptography==42.0.4
sqlparse~=0.4.2
debugpy
pandas==2.0.3
openpyxl
starrocks==1.0.6
openai==1.35.13
clickhouse-sqlalchemy==0.2.6
retrying==1.3.4
fastapi~=0.110.1
langgraph==0.2.4
langchain==0.2.14